backend/ml/sample-ml-api.py
"""
Sample ML API Implementation for Street Vendor Platform
This is a reference implementation showing how the ML APIs would work.
In production, replace with actual ML models.
"""

from flask import Flask, request, jsonify
from flask_cors import CORS
import numpy as np
import random
from datetime import datetime, timedelta
import base64
import io
from PIL import Image

app = Flask(__name__)
CORS(app)

# Sample ML model placeholders
class FootfallPredictor:
    """Mock footfall prediction model"""
    def predict_daily(self, historical_data):
        # In production, use actual trained model
        base_footfall = 150
        variation = random.randint(-30, 50)
        predictions = []
        
        for i in range(7):  # Predict next 7 days
            date = datetime.now() + timedelta(days=i+1)
            footfall = base_footfall + variation + random.randint(-20, 20)
            predictions.append({
                'date': date.strftime('%Y-%m-%d'),
                'predicted_footfall': max(50, footfall),
                'confidence': 0.75 + random.random() * 0.15
            })
        
        return predictions
    
    def predict_hourly(self, historical_data):
        # Peak hours typically 12-2 PM and 7-9 PM
        hourly_predictions = []
        peak_hours = [12, 13, 14, 19, 20, 21]
        
        for hour in range(24):
            if hour in peak_hours:
                footfall = random.randint(20, 35)
            elif 10 <= hour <= 22:
                footfall = random.randint(10, 20)
            else:
                footfall = random.randint(0, 5)
            
            hourly_predictions.append({
                'hour': hour,
                'predicted_footfall': footfall,
                'is_peak': hour in peak_hours
            })
        
        return hourly_predictions

class DemandPredictor:
    """Mock demand prediction model"""
    def predict_items(self, historical_data):
        # Sample popular items
        items = [
            {'name': 'Samosa', 'predicted_demand': 150, 'trend': 'increasing'},
            {'name': 'Tea', 'predicted_demand': 200, 'trend': 'stable'},
            {'name': 'Vada Pav', 'predicted_demand': 100, 'trend': 'increasing'},
            {'name': 'Coffee', 'predicted_demand': 80, 'trend': 'decreasing'},
            {'name': 'Pakora', 'predicted_demand': 60, 'trend': 'seasonal'}
        ]
        
        return {
            'predictions': items,
            'recommendation': 'Stock up on Samosa and Tea for tomorrow',
            'confidence': 0.82
        }

class HygieneAnalyzer:
    """Mock hygiene analysis model"""
    def analyze_image(self, image_base64):
        # In production, use computer vision model
        # For demo, return random scores
        
        overall_score = random.randint(70, 95)
        
        return {
            'overallScore': overall_score,
            'details': {
                'cleanliness': {
                    'score': random.randint(70, 95),
                    'observations': [
                        'Clean cooking surfaces detected',
                        'Proper food storage observed'
                    ]
                },
                'foodSafety': {
                    'score': random.randint(65, 90),
                    'observations': [
                        'Food items properly covered',
                        'Temperature control measures in place'
                    ]
                },
                'personalHygiene': {
                    'score': random.randint(70, 95),
                    'observations': [
                        'Staff wearing clean attire',
                        'Hand washing facilities available'
                    ]
                },
                'storageConditions': {
                    'score': random.randint(70, 90),
                    'observations': [
                        'Organized storage area',
                        'FIFO system implemented'
                    ]
                }
            },
            'confidence': 0.88,
            'modelVersion': '2.1',
            'recommendations': [
                'Maintain current cleanliness standards',
                'Consider adding hand sanitizer dispensers',
                'Regular deep cleaning recommended'
            ]
        }

# Initialize models
footfall_model = FootfallPredictor()
demand_model = DemandPredictor()
hygiene_model = HygieneAnalyzer()

# API Routes
@app.route('/health', methods=['GET'])
def health_check():
    return jsonify({
        'status': 'healthy',
        'models': {
            'footfall': 'loaded',
            'demand': 'loaded',
            'hygiene': 'loaded'
        },
        'version': '1.0'
    })

@app.route('/predict/footfall/daily', methods=['POST'])
def predict_daily_footfall():
    data = request.json
    predictions = footfall_model.predict_daily(data)
    
    return jsonify({
        'success': True,
        'predictions': predictions,
        'model_version': 'v1.2'
    })

@app.route('/predict/footfall/hourly', methods=['POST'])
def predict_hourly_footfall():
    data = request.json
    predictions = footfall_model.predict_hourly(data)
    
    return jsonify({
        'success': True,
        'predictions': predictions,
        'model_version': 'v1.2'
    })

@app.route('/predict/items', methods=['POST'])
def predict_item_demand():
    data = request.json
    predictions = demand_model.predict_items(data)
    
    return jsonify({
        'success': True,
        'data': predictions,
        'model_version': 'v1.0'
    })

@app.route('/hygiene/check', methods=['POST'])
def check_hygiene():
    data = request.json
    
    if 'image' not in data:
        return jsonify({'error': 'No image provided'}), 400
    
    result = hygiene_model.analyze_image(data['image'])
    
    return jsonify({
        'success': True,
        **result
    })

if __name__ == '__main__':
    print("ML API Server starting on port 8000...")
    app.run(host='0.0.0.0', port=8000, debug=True)