backend/ml/requirements.txt
flask==2.3.3
flask-cors==4.0.0
numpy==1.24.3
pillow==10.0.0
scikit-learn==1.3.0
pandas==2.0.3
tensorflow==2.13.0
opencv-python==4.8.0.76
python-dotenv==1.0.0